pyproject.toml
[project]
name = "tfg-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.28.0",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28", 
    "langgraph>=0.6.3",
    "requests>=2.32.4",
]
